# backend/src/schemas/generated/types/adminNotification.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bin Al-Zain Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminNotificationType } from './adminNotificationType';

export interface AdminNotification {
  id: number;
  type: AdminNotificationType;
  title: string;
  body: string;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  createdAt: string;
}