# backend/src/schemas/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bin Al-Zain Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  items: OrderItem[];
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}