# backend/src/schemas/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bin Al-Zain Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface OrderInput {
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  items: OrderItem[];
  totalPrice: number;
  /** @nullable */
  notes?: string | null;
}