# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bin Al-Zain Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
            enum: ["pending", "confirmed", "completed", "cancelled", null]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (called before WhatsApp redirect)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for logged-in customer (broadcast + targeted)
      responses:
        "200":
          description: List of notifications with read status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: Not authenticated
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          description: Not authenticated
  /notifications/read/{id}:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /admin/notifications:
    get:
      operationId: listAdminNotifications
      tags: [notifications]
      summary: List all notifications (admin)
      responses:
        "200":
          description: All notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminNotification"
        "401":
          description: Not authenticated
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create and send a notification (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNotification"
        "401":
          description: Not authenticated
  /admin/notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [productId, nameAr, price, quantity]
      properties:
        productId:
          type: integer
        nameAr:
          type: string
        price:
          type: number
        quantity:
          type: integer
        selectedWeight:
          type: ["number", "null"]
        lineTotal:
          type: number
    Order:
      type: object
      required: [id, items, totalPrice, status, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [items, totalPrice]
      properties:
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        notes:
          type: ["string", "null"]
...
    Notification:
      type: object
      required: [id, type, title, body, createdAt, isRead]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [offer, discount, warning]
        title:
          type: string
        body:
          type: string
        customerId:
          type: ["integer", "null"]
        createdAt:
          type: string
        isRead:
          type: boolean
    AdminNotification:
      type: object
      required: [id, type, title, body, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [offer, discount, warning]
        title:
          type: string
        body:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        createdAt:
          type: string
    CreateNotificationInput:
      type: object
      required: [type, title, body]
      properties:
        type:
          type: string
          enum: [offer, discount, warning]
        title:
          type: string
        body:
          type: string
        customerId:
          type: ["integer", "null"]
...